Abstract of the project
Abstract:

The aim of the project is to introduce smart manufacturing into the metal casting foundry industry.
The specific aim is to manufacture small, intricate parts which are required only in small quantities by leveraging state-of-art technologies of 3D printing of the pattern in plastic, designing a sand mould around the pattern, removing the pattern and melting and pouring of metal. The process will be digitally controlled, analytics of data generated at every stage will help improve the process.
In this project we aim to acquire and analyse temperature data from the casting process of a simple shaped aluminium casting.
This can only be achieved by an interdisciplinary team consisting of the mechanical/metallurgy engineer for the foundry trial, electronics and software engineers.